requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
db-dtypes==1.2.0
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
executing==2.0.1
fastapi==0.110.1
Flask==3.0.3
flask-sock==0.7.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
gotrue==2.4.2
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kaleido==0.2.1
langchain-core==0.1.42
langsmith==0.1.47
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.10.0
orjson==3.10.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
postgrest==0.16.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
#pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
realtime==1.0.4
requests==2.31.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
tabulate==0.9.0
tenacity==8.2.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
twilio==9.0.5
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
vanna==0.5.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
